better legibility with example in one file
